Create README.md
This project is for research for a masters project. 
The purpose is to have a feed which looks similar to instagram. 
The participant enters their name, and then their name will appear on certain images on the feed, overlaying on the images. 

Still in construction, if you can provide any help, that would be unbelievable. 
(I have never coded before, my subject is cognition)